v1

+++ server/controllers/projects.controller.js
@@ -21,27 +21,42 @@ const getProjectById = async (req, res) => {
 
 const addProject = async (req, res) => {
   try {
-    const newProject = new Project(req.body);
-    const savedProject = await newProject.save();
-    return res.status(201).json(savedProject);
+    if (req.admin.role === "admin") {
+      const newProject = new Project(req.body);
+      const savedProject = await newProject.save();
+      return res.status(201).json(savedProject);
+    } else if (req.admin.role === "tester") {
+      return res.status(200).json({
+        message: "Testing Account : Cannot add projects!",
+      });
+    }
   } catch (error) {
-    return res.status(500).json({ message: error.message });
+    return res.status(500).json({
+      message: error.message,
+    });
   }
 };
 
 const updateProject = async (req, res) => {
   try {
-    const updatedProject = await Project.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      {
-        new: true,
-        runValidators: true,
-      }
-    );
-    if (!updatedProject)
-      return res.status(404).json({ message: "Project not found" });
-    return res.status(200).json(updatedProject);
+    if (req.admin.role === "admin") {
+      const updatedProject = await Project.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        {
+          new: true,
+          runValidators: true,
+        }
+      );
+      if (!updatedProject)
+        return res.status(404).json({ message: "Project not found" });
+
+      return res.status(200).json({ updatedProject });
+    } else if (req.admin.role === "tester") {
+      return res.status(200).json({
+        message: "Testing Account : Cannot update projects!",
+      });
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -49,10 +64,16 @@ const updateProject = async (req, res) => {
 
 const deleteProject = async (req, res) => {
   try {
-    const deletedProject = await Project.findByIdAndDelete(req.params.id);
-    if (!deletedProject)
-      return res.status(404).json({ message: "Project not found" });
-    return res.status(200).json({ message: "Project deleted successfully" });
+    if (req.admin.role === "admin") {
+      const deletedProject = await Project.findByIdAndDelete(req.params.id);
+      if (!deletedProject)
+        return res.status(404).json({ message: "Project not found" });
+      return res.status(200).json({ message: "Project deleted successfully" });
+    } else if (req.admin.role === "tester") {
+      return res.status(200).json({
+        message: "Testing Account : Cannot dalete projects!",
+      });
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
